// src/theme/theme.d.ts
import { PaletteColor, PaletteColorOptions, createTheme } from "@mui/material";
import { blue, green, red } from "@mui/material/colors";

//incomeColorがエラーになるので型の拡張
declare module '@mui/material/styles' {
  interface Palette {
    incomeColor: PaletteColor;
    expensColor: PaletteColor;
    balanceColor: PaletteColor;
  }
  //パレットの初期設定
  interface PaletteOptions {
    incomeColor?: PaletteColorOptions;
    expensColor?: PaletteColorOptions;
    balanceColor?: PaletteColorOptions;
  }
}

export const theme = createTheme({
  typography: {
    fontFamily: 'Noto Sans JP, Roboto, "Helvetica Neue", Arial, sans-serif',
    fontWeightRegular: 400,
    fontWeightMedium: 500,
    fontWeightBold: 700
  },

  palette: {
    incomeColor: {
      main: blue[500],
      light: blue[100],
      dark: blue[700]
    },
    expensColor: {
      main: red[500],
      light: red[100],
      dark: red[700]
    },
    balanceColor: {
      main: green[500],
      light: green[100],
      dark: green[700]
    }
  }
})